+++ packages/design-system-shared/src/types/FilterButton/index.ts
@@ -1,0 +1,4 @@
+export {
+  FilterButtonVariant,
+  type FilterButtonPropsShared,
+} from './FilterButton.types';