+++ packages/design-system-shared/src/types/ListItemMultiSelect/index.ts
@@ -1,0 +1,1 @@
+export { type ListItemMultiSelectPropsShared } from './ListItemMultiSelect.types';